DEL-647 - Represent number as decimal
I also bumped the version to 1.1.0.

+++ KenticoCloud.Deliver/Models/Item/ContentItem.cs
@@ -58,9 +58,9 @@ namespace KenticoCloud.Deliver
         /// Gets a number value from an element.
         /// </summary>
         /// <param name="element">Element name.</param>
-        public double GetNumber(string element)
+        public decimal GetNumber(string element)
         {
-            return GetElementValue<double>(element);
+            return GetElementValue<decimal>(element);
         }
 
         /// <summary>
